I think Lab7 framework is now ready for review. Don't do a pull request because there's working code in the commit history

+++ labs/lab7/src/test/java/edu/indiana/cs/c212/Lab7Test.java
@@ -123,7 +123,7 @@ public class Lab7Test {
 		BeatlesGreatestHitsTracks.add(hereComesTheSun);
 		BeatlesGreatestHitsTracks.add(heyJude);
 		CD beatles = new CD("2 Greatest Hits", "The Beatles", 1982, BeatlesGreatestHitsTracks, "Capitol Records", "British Rock");
-		assertEquals(2, beatles.getNumberOfTracks());
+		assertEquals(2, (int)beatles.getNumberOfTracks());
 	}
 	
 	@Test
